removed warnings
git-svn-id: http://svn.seqan.de/seqan/trunk@5328 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/tests/basic/test_basic.cpp
@@ -331,7 +331,7 @@ void Test_Iter()
 	it3++;
 	SEQAN_ASSERT_EQ(*it3, 'd');
 
-	SEQAN_ASSERT_EQ(position(it3), 3);
+	SEQAN_ASSERT_EQ((int)position(it3), 3);
 
 	setPosition(it3, 2);
 	SEQAN_ASSERT_EQ(*it3, 'c');
